src/app/components/authentication/components/user-registration/user-registration.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { UserRegistrationComponent } from './user-registration.component';

describe('Component: UserRegistration', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [UserRegistrationComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([UserRegistrationComponent],
      (component: UserRegistrationComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(UserRegistrationComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(UserRegistrationComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-user-registration></app-user-registration>
  `,
  directives: [UserRegistrationComponent]
})
class UserRegistrationComponentTestController {
}

